fix weird quicklens bug that causes "error while emitting _".
the contents of this commit were copied from the cliche codebase where we encountered
the same problem after migrating to scala2.13. it is complete nonsense but that's what
it is.

it is like that because of some weird macro thing I don't understand -- and now I couldn't
find the github issue on the quicklens repository I had found about it at the first time
I encountered this bug either.

anyway.