Prevent mnemonic words from being copied to clipboard on tap.
Currently, if a chip does not explicitly set an onTap callback, or have
shareText explicitly set, the chip will copy its contents to the clipboard on
tap. Because it is possible that a user accidentally taps the chips presenting
the mnemonic words, this behavior should be inhibited, as the text contains both
the mnemonic word and its position in the recovery phrase.